refactor: centralize logging and CoVe imports, add logging_utils
- Move success_message, save_log, etc. from utils to logging_utils
- Remove redundant imports in commands using CoVe verification
- Add new logging_utils.py module for centralized logging helpers
- Update tests to reflect import changes
