+++ tests/unit/test_real_functionality.py
@@ -38,14 +38,14 @@ class TestActualFunctionality:
 
     def test_save_log_creates_file(self):
         """Test save_log creates actual files."""
-        from litassist.utils import save_log
+        from litassist.logging_utils import save_log
         import os
         import tempfile
 
         # Create a temp directory
         with tempfile.TemporaryDirectory() as temp_dir:
             # Patch LOG_DIR to use temp directory
-            with patch("litassist.utils.LOG_DIR", temp_dir):
+            with patch("litassist.logging_utils.LOG_DIR", temp_dir):
                 payload = {"input": "test", "response": "result"}
                 save_log("test", payload)
